Many Fixes & Picture Work & Code Beautifuler

+++ modules/admin/controllers/PartController.php
@@ -8,10 +8,8 @@ use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii\helpers\ArrayHelper;
-use yii\web\UploadedFile;
 use app\traits\FindModelTrait;
 use app\models\Part;
-use app\models\UploadForm;
 use app\models\Producer;
 use app\models\PartPicture;
 use app\models\Offer;
@@ -19,16 +17,10 @@ use app\models\Certificate;
 use app\models\Store;
 use app\models\PartCertificate;
 
-/**
- * PartController implements the CRUD actions for Part model.
- */
 class PartController extends Controller
 {
     use FindModelTrait;
 
-    /**
-     * {@inheritdoc}
-     */
     public function behaviors()
     {
         return [
@@ -41,10 +33,6 @@ class PartController extends Controller
         ];
     }
 
-    /**
-     * Lists all Part models.
-     * @return mixed
-     */
     public function actionIndex()
     {
         $dataProvider = new ActiveDataProvider([
@@ -56,12 +44,6 @@ class PartController extends Controller
         ]);
     }
 
-    /**
-     * Displays a single Part model.
-     * @param integer $id
-     * @return mixed
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     public function actionView($id)
     {
         return $this->render('view', [
@@ -69,40 +51,24 @@ class PartController extends Controller
         ]);
     }
 
-    /**
-     * Creates a new Part model.
-     * If creation is successful, the browser will be redirected to the 'view' page.
-     * @return mixed
-     */
     public function actionCreate()
     {
         $model = new Part();
-        $modelOffer = new Offer();
-        $modelOffer->scenario = 'create';
+        $modelOffer = new Offer(['scenario' => 'create']);
 
-        if ($model->load(Yii::$app->request->post())) {
-            $uploadModel = new UploadForm();
-            $uploadModel->uploadFile = UploadedFile::getInstance($model, 'picture');
-
-            if ($model->save()) {
-                $partPicture = new PartPicture();
-                $partPicture->picture = $uploadModel->upload();
-                $partPicture->part_id = $model->id;
-                $partPicture->save();
-
-                if ($certs = $model->certificate_ids) {
-                    foreach ($certs as $item) {
-                        $partCertificate = new PartCertificate();
-                        $partCertificate->part_id = $model->id;
-                        $partCertificate->certificate_id = $item;
-                        $partCertificate->save();
-                    }
+        if ($model->load(Yii::$app->request->post()) && $model->save()) {
+            if ($certs = $model->certificate_ids) {
+                foreach ($certs as $item) {
+                    $partCertificate = new PartCertificate();
+                    $partCertificate->part_id = $model->id;
+                    $partCertificate->certificate_id = $item;
+                    $partCertificate->save();
                 }
+            }
 
-                if ($modelOffer->load(Yii::$app->request->post())) {
-                    $modelOffer->part_id = $model->id;
-                    if ($modelOffer->save()) return $this->redirect(['view', 'id' => $model->id]);
-                }
+            if ($modelOffer->load(Yii::$app->request->post())) {
+                $modelOffer->part_id = $model->id;
+                if ($modelOffer->save()) return $this->redirect(['view', 'id' => $model->id]);
             }
         }
 
@@ -115,44 +81,27 @@ class PartController extends Controller
         ]);
     }
 
-    /**
-     * Updates an existing Part model.
-     * If update is successful, the browser will be redirected to the 'view' page.
-     * @param integer $id
-     * @return mixed
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     public function actionUpdate($id)
     {
         $model = $this->findModel(Part::class, $id);
         $modelOffer = new Offer();
 
-        if ($model->load(Yii::$app->request->post())) {
-            $uploadModel = new UploadForm();
-            $uploadModel->uploadFile = UploadedFile::getInstance($model, 'picture');
-            
-            if ($model->save()) {
-                $partPicture = new PartPicture();
-                $partPicture->picture = $uploadModel->upload();
-                $partPicture->part_id = $model->id;
-                $partPicture->save();
-
-                if ($certs = $model->certificate_ids) {
-                    PartCertificate::deleteAll(['part_id' => $model->id]);
-                    foreach ($certs as $item) {
-                        $partCertificate = new PartCertificate();
-                        $partCertificate->part_id = $model->id;
-                        $partCertificate->certificate_id = $item;
-                        $partCertificate->save();
-                    }
+        if ($model->load(Yii::$app->request->post()) && $model->save()) {
+            if ($certs = $model->certificate_ids) {
+                PartCertificate::deleteAll(['part_id' => $model->id]);
+                foreach ($certs as $item) {
+                    $partCertificate = new PartCertificate();
+                    $partCertificate->part_id = $model->id;
+                    $partCertificate->certificate_id = $item;
+                    $partCertificate->save();
                 }
-
-                if ($modelOffer->load(Yii::$app->request->post())) {
-                    $modelOffer->part_id = $model->id;
-                    if ($modelOffer->price && $modelOffer->store_id) $modelOffer->save();
-                }
-                return $this->redirect(['view', 'id' => $model->id]);
             }
+
+            if ($modelOffer->load(Yii::$app->request->post())) {
+                $modelOffer->part_id = $model->id;
+                if ($modelOffer->price && $modelOffer->store_id) $modelOffer->save();
+            }
+            return $this->redirect(['view', 'id' => $model->id]);
         }
 
         return $this->render('update', [
@@ -165,13 +114,6 @@ class PartController extends Controller
         ]);
     }
 
-    /**
-     * Deletes an existing Part model.
-     * If deletion is successful, the browser will be redirected to the 'index' page.
-     * @param integer $id
-     * @return mixed
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     public function actionDelete($id)
     {
         $this->findModel(Part::class, $id)->delete();
